pyproject.toml
[tool.poetry]
name = "postgresqlmygithup"
version = "0.1.0"
description = ""
authors = ["ZhenyaKasatkina <knopisha.zh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
